feature: running tasks progress
