+++ packages/monitor/src/server.ts
@@ -1,14 +1,13 @@
 import { Gauge, Histogram, Counter } from 'prom-client';
 import Logger from 'bunyan';
 
-import * as dotenv from 'dotenv';
-dotenv.config()
 
 import { NomadContext } from '@nomad-xyz/sdk';
 
 import { register } from 'prom-client';
 import express, { Response } from 'express';
 import { TaskRunner } from './taskRunner';
+import { sleep } from './utils';
 export const prefix = `nomad_metrics`;
 
 export enum MessageStages {
@@ -187,72 +186,4 @@ export class MonitoringCollector extends MetricsCollector {
   incGoldskyErrors(query: GoldSkyQuery, code: string) {
     this.goldskyErrors.labels(code, query, this.environment).inc();
   }
-}
-
-export class HomeStatusCollector extends TaskRunner {
-  ctx: NomadContext;
-  logger: Logger;
-  metrics: MonitoringCollector;
-
-  constructor(
-    ctx: NomadContext,
-    logger: Logger,
-    metrics: MonitoringCollector,
-  ) {
-    super();
-    this.ctx = ctx;
-    this.logger = logger;
-    this.metrics = metrics;
-  }
-
-  async runTasks() {
-    await Promise.all([
-      this.checkAllHomes()
-    ])
-  }
-
-  async checkAllHomes(): Promise<void> {
-    await this.checkHomes(this.ctx.domainNumbers);
-  }
-
-  async checkHomes(networks: (string | number)[]): Promise<void> {
-    for (const n of networks) {
-      await this.checkHome(n);
-    }
-  }
-
-  async checkHome(nameOrDomain: string | number): Promise<void> {
-    const domain = this.ctx.resolveDomain(nameOrDomain);
-    const home = this.ctx.mustGetCore(domain).home;
-    console.log(`Check home ${nameOrDomain}`);
-    const state = await home.state();
-    console.log(`Got home ${nameOrDomain}`);
-    if (state === 2) {
-      this.metrics.setHomeState(nameOrDomain.toString(), true);
-    } else {
-      this.metrics.setHomeState(nameOrDomain.toString(), false);
-    }
-  }
-
-  // async healthy(): Promise<boolean> {
-  //   try {
-  //     const state = await this.home.state();
-  //     if (state !== 1) {
-  //       return false;
-  //     } else {
-  //       return true;
-  //     }
-  //   } catch (e: any) {
-  //     this.logger.warn(
-  //       `Couldn't collect home state for ${this.domain} domain. Error: ${e.message}`,
-  //     );
-  //     throw new Error()
-  //     // TODO! something
-  //   }
-  //   return true; // BAD!!!
-  // }
-
-  // get failed(): boolean {
-  //   return !this.healthy;
-  // }
 }
