+++ packages/monitor/src/utils.ts
@@ -21,3 +21,5 @@ export enum BunyanLevel {
       environment: environment,
     });
   }
+
+  export const sleep = (ms: number) => new Promise((res) => setTimeout(res, ms));
